% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckCoverage.R
\name{CheckCellCoverage}
\alias{CheckCellCoverage}
\title{CheckCellCoverage}
\usage{
CheckCellCoverage(so1, so2)
}
\description{
This function take 2 Seurat3 Object as input, and return the intersection of two object by Cell ID.

CheckCellCoverage()
}
