% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterDifference.R
\name{ClusterDifference}
\alias{ClusterDifference}
\title{compare two cluster between two different seurat object}
\usage{
ClusterDifference(so1, c1, so2, c2)
}
\value{
a list of genes
ClusterDifference()
}
\description{
This function take 2 seurat object, and their corresponding clusternumber as input, return a set of genes that make the difference
}
