% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeConfusionMatrix.R
\name{MakeConfusionMatrix}
\alias{MakeConfusionMatrix}
\title{Make a confusion matrix}
\usage{
MakeConfusionMatrix(so1, so2)
}
\value{
a dataframe contain the confusion matrix
MakeConfusionMatrix()
}
\description{
This function take 2 seurat object as input, return a confusion matrix
}
